pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "docx2txt>=0.9",
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.0",
    "langchain-pinecone>=0.2.8",
    "langchainhub>=0.1.21",
    "markdown>=3.8.2",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
    "sentence-transformers>=5.0.0",
    "sqlalchemy>=2.0.41",
    "unstructured>=0.18.9",
]
